# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FACTA — Independent product intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meal-logs:
    get:
      operationId: listMealLogs
      tags: [mealLogs]
      summary: List meal logs for a session and date
      parameters:
        - name: session_id
          in: query
          required: true
          schema: { type: string }
        - name: date_str
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Meal logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealLog"
    post:
      operationId: addMealLog
      tags: [mealLogs]
      summary: Log a product to a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealLogInput"
      responses:
        "201":
          description: Created meal log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLog"
  /meal-logs/{id}:
    delete:
      operationId: deleteMealLog
      tags: [mealLogs]
      summary: Delete a meal log entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }

  # ── Search ────────────────────────────────────────────────────────────────
...
  /guides:
    get:
      operationId: listGuides
      tags: [guides]
      summary: List published FACTA Guides
      parameters:
        - name: goal_slug
          in: query
          schema: { type: ["string", "null"] }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Guides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuideSummary"
  /guides/{slug}:
    get:
      operationId: getGuide
      tags: [guides]
      summary: Get a FACTA Guide by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuideDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Collections ───────────────────────────────────────────────────────────
...
  /admin/guides:
    post:
      operationId: adminCreateGuide
      tags: [admin]
      summary: Create a guide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGuideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuideSummary"
  /admin/guides/{id}:
    patch:
      operationId: adminUpdateGuide
      tags: [admin]
      summary: Update a guide
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGuideInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuideSummary"
...
  /products/{id}/news:
    get:
      operationId: getProductNews
      tags: [products]
      summary: AI web search for brand/product news with sentiment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: News intelligence for this product's brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductNewsResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    NewsArticle:
      type: object
      required: [title]
      properties:
        title: { type: string }
        url: { type: ["string", "null"] }
        reportType:
          type: string
          enum: [news, advertorial, press_release, unknown]
    ProductNewsResponse:
      type: object
      required: [sentiment, articles]
      properties:
        sentiment:
          type: string
          enum: [positive, negative, mixed, neutral, none]
        summary: { type: ["string", "null"] }
        summaryZh: { type: ["string", "null"] }
        articles:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        fetchedAt: { type: ["string", "null"] }

    # ── Core entities ─────────────────────────────────────────────────────
...
    MealLogInput:
      type: object
      required: [sessionId, productId, mealType, dateStr]
      properties:
        sessionId: { type: string }
        productId: { type: integer }
        mealType:  { type: string }
        dateStr:   { type: string }
        note:      { type: ["string", "null"] }
    MealLog:
      type: object
      required: [id, sessionId, productId, mealType, dateStr, loggedAt]
      properties:
        id:          { type: integer }
        sessionId:   { type: string }
        productId:   { type: integer }
        mealType:    { type: string }
        dateStr:     { type: string }
        loggedAt:    { type: string }
        note:        { type: ["string", "null"] }
        productName: { type: ["string", "null"] }
        productNameZh: { type: ["string", "null"] }
        imageUrl:    { type: ["string", "null"] }
        overallScore: { type: ["integer", "null"] }
        scoreGrade:  { type: ["string", "null"] }

    # ── Search ────────────────────────────────────────────────────────────
...
    GuideSummary:
      type: object
      required: [id, slug, titleZh, status]
      properties:
        id:          { type: integer }
        slug:        { type: string }
        title:       { type: string }
        titleZh:     { type: string }
        summaryZh:   { type: ["string", "null"] }
        summary:     { type: ["string", "null"] }
        goalId:      { type: ["integer", "null"] }
        coverImageUrl: { type: ["string", "null"] }
        status:      { type: string }
        publishedAt: { type: ["string", "null"] }
    GuideDetail:
      allOf:
        - $ref: "#/components/schemas/GuideSummary"
        - type: object
          properties:
            bodyZh:      { type: ["string", "null"] }
            body:        { type: ["string", "null"] }
            limitationsZh: { type: ["string", "null"] }
            limitations: { type: ["string", "null"] }
            sources:
              type: array
              items:
                type: object
                properties:
                  citation:      { type: string }
                  url:           { type: ["string", "null"] }
                  publishedYear: { type: ["integer", "null"] }
            evidenceLastReviewedAt: { type: ["string", "null"] }
            reviewDueDate:          { type: ["string", "null"] }

    # ── Collections ───────────────────────────────────────────────────────
...
    AdminGuideInput:
      type: object
      properties:
        slug:          { type: string }
        title:         { type: string }
        titleZh:       { type: string }
        summaryZh:     { type: ["string", "null"] }
        body:          { type: ["string", "null"] }
        bodyZh:        { type: ["string", "null"] }
        goalId:        { type: ["integer", "null"] }
        status:        { type: string }
...