# lib/api-zod/src/generated/types/mealLogInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FACTA — Independent product intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface MealLogInput {
  sessionId: string;
  productId: number;
  mealType: string;
  dateStr: string;
  /** @nullable */
  note?: string | null;
}